Some chages, new comments, and a BUG in the Composite class was fixed (the one that throws an exception about in the matrix concatenation) in the method (get_matrix_train and test).
A new file (experiment_advanced) has been added in order to develop a
new script that represents the test instances waiting in line for input
user. This is in order to ingrate the framework with weka crossfold
validation. (Currently it is just a copy of the simple experiment, I
will code it later :) )